+++ scripts/test_sudo.sh
@@ -1,0 +1,4 @@
+echo "Testing sudo..."
+sudo ls /
+
+echo "DONE"